use nr l for longs

+++ src/test/java/com/google/sps/BookmarkPlaceGuideServletTest.java
@@ -131,23 +131,15 @@ public final class BookmarkPlaceGuideServletTest {
   private static final String EMAIL = "user@gmail.com";
   private static final Set<Long> EMPTY_BOOKMARKED_PLACE_GUIDES_IDS = null;
   private static final Set<Long> BOOKMARKED_PLACE_GUIDES_IDS_A =
-      new HashSet<>(
-          Arrays.asList(
-              (long) 1, (long) 2, (long) 3, (long) 4, (long) 5, (long) 6, (long) 7, (long) 8,
-              (long) 9, (long) 10));
+      new HashSet<>(Arrays.asList(1l, 2l, 3l, 4l, 5l, 6l, 7l, 8l, 9l, 10l));
   private static final Set<Long> BOOKMARKED_PLACE_GUIDES_IDS_B =
       new HashSet<>(
           Arrays.asList(
-              (long) 1, (long) 2, (long) 3, (long) 4, (long) 5, (long) 6, (long) 7, (long) 8,
-              (long) 9, (long) 10, (long) 11, (long) 12, (long) 13, (long) 14, (long) 15, (long) 16,
-              (long) 17, (long) 18, (long) 19, (long) 20, (long) 21, (long) 22, (long) 23,
-              (long) 24, (long) 25));
+              1l, 2l, 3l, 4l, 5l, 6l, 7l, 8l, 9l, 10l, 11l, 12l, 13l, 14l, 15l, 16l, 17l, 18l, 19l,
+              20l, 21l, 22l, 23l, 24l, 25l));
   private static final Set<Long> BOOKMARKED_PLACE_GUIDES_IDS_C =
       new HashSet<>(
-          Arrays.asList(
-              (long) 1, (long) 2, (long) 3, (long) 4, (long) 5, (long) 6, (long) 7, (long) 8,
-              (long) 9, (long) 10, (long) 11, (long) 12, (long) 13, (long) 14, (long) 15,
-              (long) 26));
+          Arrays.asList(1l, 2l, 3l, 4l, 5l, 6l, 7l, 8l, 9l, 10l, 11l, 12l, 13l, 14l, 15l, 26l));
   private static final String NAME = "username";
   private static final String SELF_INTRODUCTION = "I am the user";
   private static final String IMG_KEY = "/img.com";
